% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{accuracy_g}
\alias{accuracy_g}
\title{Calculate the accuracy of the approximation of g}
\usage{
accuracy_g(n, p, R2, c)
}
\arguments{
\item{n}{The number of observations}

\item{p}{The number of covariates}

\item{R2}{The correlation co-efficient, squared}

\item{c}{The c parameter}
}
\value{
The accuracy of the approximation of g
}
\description{
Calculate the accuracy of the approximation of g
}

