% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{accuracy_sigma2}
\alias{accuracy_sigma2}
\title{Calculate the accuracy of the approximation of sigma2}
\usage{
accuracy_sigma2(n, p, R2, r, s)
}
\arguments{
\item{n}{The number of observations}

\item{p}{The number of covariates}

\item{R2}{The correlation co-efficient, squared}

\item{r}{The first parameter of the q(sigma2) distribution}

\item{s}{The second parameter of the q(sigma2) distribution}
}
\value{
The accuracy of the approximation of sigma2
}
\description{
Calculate the accuracy of the approximation of sigma2
}

