% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aic}
\alias{aic}
\title{Calculate AIC}
\usage{
aic(n, p, R2)
}
\arguments{
\item{n}{The number of observations}

\item{p}{The number of covariates}

\item{R2}{The correlation co-efficient, squared}
}
\value{
The value of the marginal log-likelihood log p(y)
}
\description{
Calculate AIC
}

