% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (Mark Greenaway's conflicted copy 2016-10-21).R, R/RcppExports.R
\name{E_g_one_plus_g_squared}
\alias{E_g_one_plus_g_squared}
\title{Calculate the second integral for the posterior variance of beta_hat}
\usage{
E_g_one_plus_g_squared(n, p, R2)

E_g_one_plus_g_squared(n, p, R2)
}
\arguments{
\item{n}{The number of samples}

\item{p}{The number of covariates}

\item{R2}{The correlation}

\item{n}{The number of samples}

\item{p}{The number of covariates}

\item{R2}{The correlation}
}
\value{
The value of the integral

The value of the integral
}
\description{
Calculate the second integral for the posterior variance of beta_hat

Calculate the second integral for the posterior variance of beta_hat
}

