% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{E_g_y}
\alias{E_g_y}
\title{Calculate the exact posterior expectation of g}
\usage{
E_g_y(n, p, R2)
}
\arguments{
\item{n}{The number of observations}

\item{p}{The number of covariates}

\item{R2}{The correlation coefficient squared}
}
\value{
The exact posterior expectation of g
}
\description{
Calculate the exact posterior expectation of g
}

