% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{E_q_g_c}
\alias{E_q_g_c}
\title{Calculate the approximate posterior expectation of g}
\usage{
E_q_g_c(n, p, R2, c)
}
\arguments{
\item{n}{The number of observations}

\item{p}{The number of covariates}

\item{R2}{The correlation coefficient squared}

\item{c}{The co-efficient c of the exponential function exp(-c/g)}
}
\value{
The approximate posterior expectation of g
}
\description{
Calculate the approximate posterior expectation of g
}

