% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (Mark Greenaway's conflicted copy 2016-10-21).R, R/RcppExports.R
\name{elbo}
\alias{elbo}
\title{Calculate the variational lower bound}
\usage{
elbo(n, p, c, s, tau_g, log_det_XTX, log_det_mSigma)

elbo(n, p, c, s, tau_g, log_det_XTX, log_det_mSigma)
}
\arguments{
\item{n}{The number of observations}

\item{p}{The number of covariates}

\item{c}{The variational parameter c}

\item{s}{The variational parameter s}

\item{tau_g}{The variational parameter tau_g}

\item{log_det_XTX}{The log of the determinant of X^T X}

\item{log_det_mSigma}{The log of the determinant of mSigma}

\item{n}{The number of observations}

\item{p}{The number of covariates}

\item{c}{The variational parameter c}

\item{s}{The variational parameter s}

\item{tau_g}{The variational parameter tau_g}

\item{log_det_XTX}{The log of the determinant of X^T X}

\item{log_det_mSigma}{The log of the determinant of mSigma}
}
\value{
The variational lower bound

The variational lower bound
}
\description{
Calculate the variational lower bound

Calculate the variational lower bound
}

