% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exact_precision}
\alias{exact_precision}
\title{Calculate the exact precision}
\usage{
exact_precision(n, p, R2)
}
\arguments{
\item{n}{The number of observations}

\item{p}{The number of covariates}

\item{R2}{The correlation coefficient squared}
}
\value{
The exact precision
}
\description{
Calculate the exact precision
}

