% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_experiments.R
\name{exact_var_g}
\alias{exact_var_g}
\title{The exact variance of g/(1 + g)}
\usage{
exact_var_g(n, p, R2)
}
\arguments{
\item{n}{The number of observations}

\item{p}{The number of covariates}

\item{R2}{The correlation co-efficient squared}
}
\value{
The exact variance of g/(1 + g)
}
\description{
The exact variance of g/(1 + g)
}

