% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (Mark Greenaway's conflicted copy 2016-10-21).R, R/RcppExports.R
\name{g_ints}
\alias{g_ints}
\title{Calculate posterior variance integrals based on vectors of n, p and R2 values}
\usage{
g_ints(vn, vp, vR2)

g_ints(vn, vp, vR2)
}
\arguments{
\item{vn}{A numeric vector of n values}

\item{vp}{A numeric vector of p values}

\item{vR2}{A numeric value of}

\item{vn}{A numeric vector of n values}

\item{vp}{A numeric vector of p values}

\item{vR2}{A numeric value of}
}
\value{
List of the values of the posterior variance integrals. E_g_one_plus_g contains the result of the first
integral, while E_g_one_plus_g_squared contains the second.

List of the values of the posterior variance integrals. E_g_one_plus_g contains the result of the first
integral, while E_g_one_plus_g_squared contains the second.
}
\description{
Calculate posterior variance integrals based on vectors of n, p and R2 values

Calculate posterior variance integrals based on vectors of n, p and R2 values
}

