% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (Mark Greenaway's conflicted copy 2016-10-21).R, R/RcppExports.R
\name{log_p}
\alias{log_p}
\title{Calculate marginal log-likelihood log p(y)}
\usage{
log_p(n, p, R2)

log_p(n, p, R2)
}
\arguments{
\item{n}{The number of observations}

\item{p}{The number of covariates}

\item{R2}{The correlation co-efficient, squared}

\item{n}{The number of observations}

\item{p}{The number of covariates}

\item{R2}{The correlation co-efficient, squared}
}
\value{
The value of the marginal log-likelihood log p(y)

The value of the marginal log-likelihood log p(y)
}
\description{
Calculate marginal log-likelihood log p(y)

Calculate marginal log-likelihood log p(y)
}

