% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_sigma2_y}
\alias{p_sigma2_y}
\title{p_sigma2_y}
\usage{
p_sigma2_y(n, p, R2, sigma2)
}
\arguments{
\item{n}{The number of observations}

\item{p}{The number of covariates}

\item{R2}{The correlation co-efficient, squared}

\item{sigma2}{The value of sigma2}
}
\value{
The value of p_sigma2_y
}
\description{
p_sigma2_y
}

