% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graphs.R
\name{plot_graphs}
\alias{plot_graphs}
\title{Plot graphs of the log-lower bound of tau_g for various values of n, p and R2}
\usage{
plot_graphs()
}
\description{
Plot graphs of the log-lower bound of tau_g for various values of n, p and R2
}

