% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{q_sigma2}
\alias{q_sigma2}
\title{q_sigma2}
\usage{
q_sigma2(r, s, sigma2)
}
\arguments{
\item{r}{The first parameter of the Inverse Gamma distribution}

\item{s}{The second parameter of the Inverse Gamma distribution}

\item{sigma2}{The value of sigma2}
}
\value{
The value of q_sigma2
}
\description{
q_sigma2
}

