% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (Mark Greenaway's conflicted copy 2016-10-21).R, R/RcppExports.R
\name{tau_g}
\alias{tau_g}
\title{Calculate tau_g from n, p and R2}
\usage{
tau_g(n, p, R2)

tau_g(n, p, R2)
}
\arguments{
\item{n}{The number of observations}

\item{p}{The number of parameters}

\item{R2}{The correlation co-efficient}

\item{n}{The number of observations}

\item{p}{The number of parameters}

\item{R2}{The correlation co-efficient}
}
\value{
The log-lower bound of tau_g

The log-lower bound of tau_g
}
\description{
Calculate tau_g from n, p and R2

Calculate tau_g from n, p and R2
}

