% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (Mark Greenaway's conflicted copy 2016-10-21).R, R/RcppExports.R
\name{taug}
\alias{taug}
\title{Calculate tau_g}
\usage{
taug(n, mGraycode, vR2, vlog_ZE)

taug(n, mGraycode, vR2, vlog_ZE)
}
\arguments{
\item{n}{The number of samples}

\item{mGraycode}{The graycode matrix}

\item{vR2}{The vector of correlations}

\item{vlog_ZE}{The vector of log(ZE)}

\item{n}{The number of samples}

\item{mGraycode}{The graycode matrix}

\item{vR2}{The vector of correlations}

\item{vlog_ZE}{The vector of log(ZE)}
}
\description{
Calculate tau_g

Calculate tau_g
}

