% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports (Mark Greenaway's conflicted copy 2016-10-21).R, R/RcppExports.R
\name{var_g_over_one_plus_g}
\alias{var_g_over_one_plus_g}
\title{Calculate variance of tau_g from n, p and R2}
\usage{
var_g_over_one_plus_g(n, p, R2)

var_g_over_one_plus_g(n, p, R2)
}
\arguments{
\item{n}{The number of observations}

\item{p}{The number of parameters}

\item{R2}{The correlation co-efficient}

\item{n}{The number of observations}

\item{p}{The number of covariates}

\item{R2}{The correlation co-efficient, squared}
}
\value{
The log-lower bound of tau_g

The variance of g/(1+g) under q
}
\description{
Calculate variance of tau_g from n, p and R2

Variance of g over one plug g under q
}

