% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{var_int2}
\alias{var_int2}
\title{Calculate the second integral for the posterior variance of beta_hat}
\usage{
var_int2(n, p, R2)
}
\arguments{
\item{n}{The number of samples}

\item{p}{The number of covariates}

\item{R2}{The correlation}
}
\value{
The value of the integral
}
\description{
Calculate the second integral for the posterior variance of beta_hat
}

