% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{var_ints}
\alias{var_ints}
\title{Calculate posterior variance integrals based on vectors of n, p and R2 values}
\usage{
var_ints(vn, vp, vR2)
}
\arguments{
\item{vn}{A numeric vector of n values}

\item{vp}{A numeric vector of p values}

\item{vR2}{A numeric value of}
}
\value{
List of the values of the posterior variance integrals. var_int1 contains the result of the first
integral, while var_int2 contains the second.
}
\description{
Calculate posterior variance integrals based on vectors of n, p and R2 values
}

